requirements.txt
fastapi==0.96.0
uvicorn==0.21.1
boto3==1.26.143
supabase==1.0.2
pydantic==1.10.6
python-dotenv==1.0.0
google-auth-oauthlib
google-api-python-client==2.82.0
itsdangerous==2.1.2
pytest==7.3.1
loguru==0.7.0